+++ tests/unit/test_db.py
@@ -1,6 +1,6 @@
 """Tests for rembrandt.db."""
 
-from datetime import datetime, timedelta
+from datetime import datetime
 
 import pytest